% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_ad.R
\name{bind_literals}
\alias{bind_literals}
\title{Bind literals}
\usage{
bind_literals(dag)
}
\arguments{
\item{dag}{to bind literals in}
}
\description{
Bind literals
}
\examples{
ast <- infer_ast("cos(2*x1 + x2) + x2^2")
dag <- make_dag(ast)
dag <- collect_leaves(dag)

if (interactive()) {
  ggdag(dag)
}

dag2 <- init_graph(dag)

if (interactive()) {
  ggdag(dag2)
}

dag2 <- bind_literals(dag2)

if (interactive()) {
  ggdag(dag2)
}

}
