% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_ad.R
\name{bind_symbols}
\alias{bind_symbols}
\title{Bind symbols}
\usage{
bind_symbols(dag, values)
}
\arguments{
\item{dag}{to bind symbols in}

\item{values}{to bind for symbols, named list}
}
\description{
Bind symbols
}
\examples{
ast <- infer_ast("cos(2*x1 + x2) + x2^2")
dag <- make_dag(ast)
dag <- collect_leaves(dag)

if (interactive()) {
  ggdag(dag)
}

dag2 <- init_graph(dag)

if (interactive()) {
  ggdag(dag2)
}

dag2 <- bind_symbols(dag2, values = list(x1 = 1, x2 = 2))

if (interactive()) {
  ggdag(dag2)
}

}
