% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_leaves.R
\name{collect_leaves}
\alias{collect_leaves}
\title{Collect symbol leaves}
\usage{
collect_leaves(dag)
}
\arguments{
\item{dag}{in which same symbol leaves should be collected}
}
\description{
Collect symbol leaves
}
\examples{
symlist <- infer_ast("cos(2*x1 + x2) + x2^2")
dag <- make_dag(symlist)
if (interactive()) {
  plot(dag)
}

dag2 <- collect_leaves(dag)
if (interactive()) {
  plot(dag2)
}

}
