% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_ad.R
\name{forward_computation}
\alias{forward_computation}
\title{Performs the function calculation forwards in graph}
\usage{
forward_computation(dag, values)
}
\arguments{
\item{dag}{to perform calculations in}

\item{values}{a named list of symbols and their values}
}
\description{
Performs the function calculation forwards in graph
}
\examples{
ast <- infer_ast("cos(2*x1 + x2) + x2^2")
dag <- make_dag(ast)
dag <- collect_leaves(dag)
dag <- forward_computation(dag, values = list(x1 = 1, x2 = 2))
dag

eval(expression(cos(2*x1 + x2) + x2^2), list(x1 = 1, x2 = 2))

}
