% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_common.R
\name{get_values}
\alias{get_values}
\title{Get AD value of node}
\usage{
get_values(dag, nodes)
}
\arguments{
\item{dag}{to look \code{nodes} up in}

\item{nodes}{to get AD value from}
}
\description{
Get AD value of node
}
