% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggdag}
\alias{ggdag}
\title{Plot DAG}
\usage{
ggdag(
  dag,
  add_node_text = TRUE,
  add_value = FALSE,
  add_deriv = FALSE,
  r = 0.3,
  arrow_size = ggplot2::unit(0.25, "cm"),
  end_cap = ggraph::circle(0.85, "cm"),
  lineheight = 0.85,
  value_prefix = "v = ",
  deriv_prefix = "d = ",
  format_args = list()
)
}
\arguments{
\item{dag}{to plot}

\item{add_node_text}{add text to nodes}

\item{add_value}{add the value to node text}

\item{add_deriv}{add the derivative to node text}

\item{r}{radius of nodes}

\item{arrow_size}{size of arrows}

\item{end_cap}{cap end to not hide arrow}

\item{lineheight}{of node text}

\item{value_prefix}{Prefix when printing AD value}

\item{deriv_prefix}{Prefix when printing AD derivate}

\item{format_args}{Passed on to \code{formatC} as dots}
}
\description{
Plot DAG
}
\examples{
ast <- infer_ast("cos(2*x1 + x2) + x2^2")
dag <- make_dag(ast)
dag <- collect_leaves(dag)

if (interactive()) {
  ggdag(dag)
}

dag <- forward_computation(dag, values = list(x1 = 1, x2 = 2))

if (interactive()) {
  ggdag(dag)
  ggdag(dag, add_value = TRUE)
  ggdag(dag, add_value = TRUE, add_deriv = TRUE)
  ggdag(dag, add_value = TRUE, add_deriv = TRUE, format_args = list(digits = 2))
}

}
