% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dag.R
\name{make_dag}
\alias{make_dag}
\title{Convert abstract syntax tree to directed acyclic graph}
\usage{
make_dag(
  x,
  graph = igraph::make_empty_graph(directed = TRUE),
  ancestor_id = NULL,
  env = new.env(parent = emptyenv())
)
}
\description{
Convert abstract syntax tree to directed acyclic graph
}
