% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_reverse.R
\name{reverse_ad}
\alias{reverse_ad}
\title{Perform reverse algorithmic differentiation}
\usage{
reverse_ad(dag, values)
}
\arguments{
\item{dag}{in which to perform reverse algorithmic differentiation}

\item{values}{a named list of symbols and their values}
}
\description{
Perform reverse algorithmic differentiation
}
\examples{
ast <- infer_ast("cos(2*x1 + x2) + x2^2")
dag <- make_dag(ast)
dag2 <- collect_leaves(dag)
dag2 <- reverse_ad(dag2, values = list(x1 = 1, x2 = 2))
dag2

}
