% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_ad.R
\name{reverse_ad}
\alias{reverse_ad}
\title{Perform reverse algorithmic differentiation}
\usage{
reverse_ad(dag, symbols = NULL)
}
\arguments{
\item{dag}{in which to perform reverse algorithmic differentiation}

\item{symbols}{which symbols to differentiate with respect to (NULL means all)}
}
\description{
Perform reverse algorithmic differentiation
}
\examples{
ast <- infer_ast("cos(2*x1 + x2)")
dag <- make_dag(ast)
ad <- reverse_ad(dag)
ad

}
